Working typescript Menu
