+++ libraries/core-react/src/Menu/index.tsx
@@ -1,21 +1,25 @@
-// @ts-nocheck
 import React from 'react'
-import { Menu as BaseMenu } from './Menu'
+import { Menu as BaseMenu, MenuProps } from './Menu'
 import { MenuItem } from './MenuItem'
 import { MenuSection } from './MenuSection'
 import { MenuProvider } from './Menu.context'
 
-const Menu = ({ children, ...rest }) => (
+export type MenuType = typeof BaseMenu & {
+  MenuItem: typeof MenuItem
+  MenuSection: typeof MenuSection
+}
+
+const MenuWrapper = ({ children, ...rest }: MenuProps) => (
   <MenuProvider>
     <BaseMenu {...rest}>{children}</BaseMenu>
   </MenuProvider>
 )
 
+const Menu = MenuWrapper as MenuType
+
 Menu.MenuItem = MenuItem
 Menu.MenuSection = MenuSection
 
-Menu.propTypes = BaseMenu.propTypes
-Menu.defaultProps = BaseMenu.defaultProps
 Menu.displayName = BaseMenu.displayName
 
 export { Menu }
